tasks are now grouped according to source instead of destination gateway
Previously all the tasks which would end up on the same slave-gateway were batched together and encrypted with the key of the source of the first of these tasks.
Now the tasks are grouped based on their source and each list of tasks is encrypted and sent individually with the key of that source.