Handled the case, when rabbit was shutdown while gateway was running
If gateway tries to push something to rabbit, and it fails, gateway tries to
reconnect to rabbit and retries pushing.
If that fails again, the task is rejected and an appropriate message is returned.
Note, that Master-Gateway will try to push the task to other organizations, and
so the most likely error message a user will see is: "Task rejected by all organizations! Code:0"